added a check for lastSync to the bind callback
Chrjs.store now triggers tiddler events for local/pending tiddlers as
well as server specific ones. It now also contains a lastSync property
which, if null, means that the tiddler (in its current form) has never
been synced (i.e. it is local).

This adds a check before retrieving the latest text of a newly updated
(by someone else on the server) to make sure that the change hasn't come
from the window we're just about to update.